requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
distlib==0.3.3
Django==3.2.9
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
filelock==3.4.0
phonenumbers==8.12.48
Pillow==8.4.0
platformdirs==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.10.0
